+++ data-avro/src/main/java/com/linkedin/data/avro/PegasusUnionToAvroRecordConvertCallback.java
@@ -12,6 +12,7 @@ import com.linkedin.data.schema.EnumDataSchema;
 import com.linkedin.data.schema.MapDataSchema;
 import com.linkedin.data.schema.Name;
 import com.linkedin.data.schema.RecordDataSchema;
+import com.linkedin.data.schema.TyperefDataSchema;
 import com.linkedin.data.schema.UnionDataSchema;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -62,6 +63,14 @@ class PegasusUnionToAvroRecordConvertCallback implements DataSchemaTraverse.Call
     {
       DataSchema fieldSchema = field.getType().getDereferencedDataSchema();
 
+      Map<String, Object>  propagatedProperties = fieldSchema.getProperties();
+      if (field.getType().getType() == DataSchema.Type.TYPEREF)
+      {
+        // If the field schema type is TypeRef,
+        // then use the merged typeref properties instead
+        propagatedProperties = ((TyperefDataSchema) field.getType()).getMergedTyperefProperties();
+      }
+
       // The conversion from Pegasus union type to an Avro record is performed when the union appears as either the
       // field's direct type or when the field's type is an array or a map whose (nested) elements is of union type.
       // This conversion ignores the default value when specified on an array or map typed field. Since the elements in
@@ -74,7 +83,7 @@ class PegasusUnionToAvroRecordConvertCallback implements DataSchemaTraverse.Call
 
       if (modifiedSchema != null)
       {
-        overrideUnionFieldSchemaAndDefault(field, modifiedSchema, modifiedDefaultValue);
+        overrideUnionFieldSchemaAndDefault(field, modifiedSchema, modifiedDefaultValue, propagatedProperties);
       }
     }
   }
@@ -178,9 +187,10 @@ class PegasusUnionToAvroRecordConvertCallback implements DataSchemaTraverse.Call
    * @param field Reference to the field whose schema and default value is being overridden
    * @param modifiedSchema The override schema to use for the specified field
    * @param modifiedDefaultValue The override default value to use for the specified field
+   * @param propagatedProperties The properties value to use for the specified field
    */
   private void overrideUnionFieldSchemaAndDefault(RecordDataSchema.Field field,
-      DataSchema modifiedSchema, Object modifiedDefaultValue)
+      DataSchema modifiedSchema, Object modifiedDefaultValue, Map<String, Object> propagatedProperties)
   {
     // Stash the field's original type and default value, so that we can use this for reverting them back after
     // the schema translation is complete. This is because we don't want the input schema to have any modifications
@@ -190,6 +200,7 @@ class PegasusUnionToAvroRecordConvertCallback implements DataSchemaTraverse.Call
 
     field.setType(modifiedSchema);
     field.setDefault(modifiedDefaultValue);
+    field.setProperties(propagatedProperties);
   }
 
   /**
@@ -251,6 +262,7 @@ class PegasusUnionToAvroRecordConvertCallback implements DataSchemaTraverse.Call
     fields.add(discriminatorField);
 
     recordDataSchema.setFields(fields, errorMessageBuilder);
+    recordDataSchema.setProperties(unionDataSchema.getProperties());
     return  recordDataSchema;
   }